% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_stats.R
\name{add_game_stats}
\alias{add_game_stats}
\title{Add game stats}
\usage{
add_game_stats(game_stats_table, player_ids, game_id)
}
\arguments{
\item{game_stats_table}{Game stats table}

\item{player_ids}{Vector of unique identifiers of players from the players table}

\item{game_id}{Unique identifier of a game from the games table}
}
\description{
Add new game (with no data) to game_stats_table
}
