% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logs.R
\name{add_log_entry}
\alias{add_log_entry}
\title{Add log entry}
\usage{
add_log_entry(game_log, player, event = events, undo = FALSE)
}
\arguments{
\item{game_log}{Game log comprised of vector of event descriptions}

\item{player}{Player display name}

\item{event}{Game event (e.g., FGM, STL, etc.)}

\item{undo}{Boolean indicating if event was undone}
}
\description{
Create and append entry to game log
}
