% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosters.R
\name{add_roster_row}
\alias{add_roster_row}
\title{Add roster row}
\usage{
add_roster_row(team_id, players_table, rosters_table)
}
\arguments{
\item{team_id}{Unique identifier of a team from the teams table}

\item{players_table}{Players table}

\item{rosters_table}{Rosters table}
}
\value{
List of updated players_table, rosters_table, and roster_view.
}
\description{
Add new row to roster view.
}
