% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binge_raw.R
\docType{data}
\name{binge_raw}
\alias{binge_raw}
\title{ALSPAC binge eating example dataset with 6 binge eating variables assessed}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(binge_raw)
}
\description{
A synthetic example dataset for binge eating items collected at age 14
in the ALSPAC cohort.
}
\examples{
data(binge_raw)
}
\keyword{datasets}
