% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binge_scoresheet.R
\docType{data}
\name{binge_scoresheet}
\alias{binge_scoresheet}
\title{ALSPAC binge eating example scoresheet - Age 14 (ccq) variables}
\format{
R data file
}
\usage{
data(binge_scoresheet)
}
\description{
An example scoresheet for binge eating items collected at age 14
in the ALSPAC cohort.
}
\examples{
data(binge_scoresheet)
raw_vars <- binge_scoresheet$raw_vars
}
\references{
The Avon Longitudinal Study of Parents and Children
(\href{http://www.bristol.ac.uk/alspac/}{ALSPAC})
}
\keyword{datasets}
