% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_game_stats}
\alias{calc_game_stats}
\title{Game statistics}
\usage{
calc_game_stats(data)
}
\arguments{
\item{data}{Dataframe that includes columns with tallied game statistics}
}
\description{
Create dataframe with tallied and calculated game statistics
}
