% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_points}
\alias{calc_points}
\title{Points}
\usage{
calc_points(FTM, FGM2, FGM3)
}
\arguments{
\item{FTM}{Free throws made}

\item{FGM2}{Two-point field goals made}

\item{FGM3}{Three-point field goals made}
}
\description{
Calculate points scored
}
