% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_shooting}
\alias{calc_shooting}
\title{Shooting percentage}
\usage{
calc_shooting(made, attempted)
}
\arguments{
\item{made}{Shots made}

\item{attempted}{Shots attempted}
}
\description{
Calculate shooting percentage
}
