% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_true_shooting}
\alias{calc_true_shooting}
\title{True shooting percentage}
\usage{
calc_true_shooting(PTS, FTA, FGA)
}
\arguments{
\item{PTS}{Points}

\item{FTA}{Free throws attempted}

\item{FGA}{Field goals attempted}
}
\description{
Calculate true shooting percentage
}
\examples{
calc_true_shooting(18, 4, 12)

}
