% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_when_vars.R
\name{case_when_vars}
\alias{case_when_vars}
\title{case_when function
\code{case_when_vars} completes case_when evaluation and returns a new variable}
\usage{
case_when_vars(raw, scoresheet)
}
\arguments{
\item{raw}{a raw data object}

\item{scoresheet}{a properly formatted scoresheet object. Key columns that must have data for a case_when function are new_var and code. Code should be formatted in accordance with
dplyr case_when input. See \href{https://dplyr.tidyverse.org/reference/case_when.html}{case_when}}
}
\value{
appends \code{new_var} column after evaluation of case_when statements in \code{code} column of scoresheet
}
\description{
case_when function
\code{case_when_vars} completes case_when evaluation and returns a new variable
}
