% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logs.R
\name{create_log_entry}
\alias{create_log_entry}
\title{Create log entry}
\usage{
create_log_entry(event = events, player, undo = FALSE)
}
\arguments{
\item{event}{Game event (e.g., FGM, STL, etc.)}

\item{player}{Player name}

\item{undo}{Boolean indicating if event was undone}
}
\description{
Create entry for appending to game log
}
