% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logs.R
\name{create_log_header}
\alias{create_log_header}
\title{Create log header}
\usage{
create_log_header(team_id, game_id, date, opponent)
}
\arguments{
\item{team_id}{Unique identifier of a team from the teams table}

\item{game_id}{Unique identifier of a game from the games table}

\item{date}{Date game was played}

\item{opponent}{Opponent of focal team}
}
\description{
Create log header for top of game log
}
