% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{create_player_name}
\alias{create_player_name}
\title{Create player name}
\usage{
create_player_name(first_name, last_name)
}
\arguments{
\item{first_name}{Vector of player first names}

\item{last_name}{Vector of player last names}
}
\description{
Create player name for display in app
}
