% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{create_player_namenum}
\alias{create_player_namenum}
\title{Create player name with number}
\usage{
create_player_namenum(first_name, last_name, number)
}
\arguments{
\item{first_name}{Player first name}

\item{last_name}{Player last name}

\item{number}{Player number}
}
\description{
Create player name with number for display in app
}
