% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{create_player_names}
\alias{create_player_names}
\title{Create player names}
\usage{
create_player_names(data)
}
\arguments{
\item{data}{Dataframe that includes columns for FirstName, LastName, and Number}
}
\description{
Create player names for display in app
}
