% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosters.R
\name{delete_roster_row}
\alias{delete_roster_row}
\title{Delete roster row}
\usage{
delete_roster_row(roster_view, roster_row, players_table, rosters_table)
}
\arguments{
\item{roster_view}{Roster view}

\item{roster_row}{Index of row to delete from roster view}

\item{players_table}{Players table}

\item{rosters_table}{Rosters table}
}
\value{
List of updated players_table, rosters_table, and roster_view.
}
\description{
Delete row from roster view and update rosters and players tables.
}
