% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{delete_teams_row}
\alias{delete_teams_row}
\title{Delete teams row}
\usage{
delete_teams_row(
  teams_table,
  teams_row,
  players_table = NULL,
  rosters_table = NULL
)
}
\arguments{
\item{teams_table}{Teams table}

\item{teams_row}{Index of row to delete from teams table}

\item{players_table}{Players table}

\item{rosters_table}{Rosters table}
}
\value{
List of updated teams_table, players_table, and rosters_table.
}
\description{
Delete row from teams table and update rosters and players tables.
}
