% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{edit_players_row}
\alias{edit_players_row}
\title{Edit players row}
\usage{
edit_players_row(players_table, row, col, value)
}
\arguments{
\item{players_table}{Players table}

\item{row}{Row index of cell to update}

\item{value}{Value to place in cell}

\item{column}{Column index of cell to update}
}
\description{
Edit row in players table
}
