% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosters.R
\name{edit_roster_row}
\alias{edit_roster_row}
\title{Edit roster row}
\usage{
edit_roster_row(
  roster_view,
  roster_row,
  roster_col,
  value,
  players_table,
  rosters_table
)
}
\arguments{
\item{roster_view}{Roster view}

\item{roster_row}{Row index of cell to update}

\item{roster_col}{Column index of cell to update}

\item{value}{Value to place in cell}

\item{players_table}{Players table}

\item{rosters_table}{Rosters table}
}
\value{
List of updated players_table, rosters_table, and roster_view.
}
\description{
Edit row in roster view
}
