% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{edit_teams_row}
\alias{edit_teams_row}
\title{Edit teams row}
\usage{
edit_teams_row(teams_table, row, col, value)
}
\arguments{
\item{teams_table}{Teams table}

\item{row}{Row index of cell to update}

\item{col}{Column index of cell to update}

\item{value}{Value to place in cell}
}
\description{
Edit row in teams table
}
