% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{efficiency}
\alias{efficiency}
\title{Efficiency}
\usage{
efficiency(PTS, REB, AST, STL, BLK, FGA, FGM, FTA, FTM, TOV)
}
\arguments{
\item{PTS}{Points}

\item{REB}{Rebounds}

\item{AST}{Assists}

\item{STL}{Steals}

\item{BLK}{Blocks}

\item{FGA}{Field goals attempted}

\item{FGM}{Field goals made}

\item{FTA}{Free throws attempted}

\item{FTM}{Free throws made}

\item{TOV}{Turnovers}
}
\description{
Calculate efficiency
}
\examples{
efficiency(15, 9, 3, 1, 0, 6, 14, 3, 4, 1)

}
