% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_at.R
\name{filter_at_vars}
\alias{filter_at_vars}
\title{Filter at function}
\usage{
filter_at_vars(raw, metadata)
}
\arguments{
\item{raw}{a raw data object}

\item{metadata}{a metadata object}
}
\value{
tibble that is filtered at specific variables
}
\description{
Filter at function
}
