% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-stats.R
\name{increment_game_stat}
\alias{increment_game_stat}
\title{Increment game stat}
\usage{
increment_game_stat(
  games_stats_table,
  player_id,
  game_id,
  stat = c("DNP", events)
)
}
\arguments{
\item{player_id}{Unique identifiers of a player from the players table}

\item{game_id}{Unique identifier of a game from the games table}

\item{stat}{Statistic (column) to be incremented}

\item{game_stats_table}{Game stats table}
}
\description{
Increment game stat in game_stats_table
}
