% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosters.R
\name{init_rosters_table}
\alias{init_rosters_table}
\title{Initialize rosters table}
\usage{
init_rosters_table()
}
\description{
Initialize rosters table with no data. The rosters file (note plural rosters)
is saved to file but a roster view is not.
}
