% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_vars.R
\name{mean_vars}
\alias{mean_vars}
\title{Mean variables using scoresheet}
\usage{
mean_vars(raw, scoresheet)
}
\arguments{
\item{raw}{A raw data object}

\item{scoresheet}{A formatted scoresheet object}
}
\value{
raw tibble with four additional columns appended:
a column which is the raw sum of items,
a _NAs column with the count of NAs among items summed,
the _NA_percent column with the percentage of NAs among the meaned variables,
and the _weighted_mean column with the mean that is weighted by the number of non-NA items included
}
\description{
Mean variables using scoresheet
}
