% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_val_lab_func.R
\name{n_val_lab_func}
\alias{n_val_lab_func}
\title{value labels function}
\usage{
n_val_lab_func(scoresheet, i)
}
\arguments{
\item{scoresheet}{A scoresheet}

\item{i}{a scoresheet row}
}
\value{
a list of named values
}
\description{
value labels function
}
