% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_var_func.R
\name{r_var_func}
\alias{r_var_func}
\title{Title}
\usage{
r_var_func(scoresheet, i)
}
\arguments{
\item{scoresheet}{A scoresheet object}

\item{i}{a scoresheet row}
}
\value{
a list of raw variables
}
\description{
Title
}
