% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_vars.R
\name{rename_vars}
\alias{rename_vars}
\title{Rename Function
Renames variables to new variable name}
\usage{
rename_vars(raw, scoresheet)
}
\arguments{
\item{raw}{Raw data}

\item{scoresheet}{scoresheet}
}
\value{
table with renamed variables
}
\description{
Rename Function
Renames variables to new variable name
}
