% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorekeep.R
\name{score}
\alias{score}
\title{score function
The score function completes all scoring for a single step of a scoresheet. operations are conducted in the following order:
\code{select}, \code{if_else}, \code{recode}, \code{sum}, \code{case_when}, \code{filter_at}}
\usage{
score(raw, scoresheet)
}
\arguments{
\item{raw}{Raw data object}

\item{scoresheet}{Formatted scoresheet object with one step selected (all operations can be preformed simultaneously.
For sequential operations, see scorekeep function)}
}
\value{
a tibble
}
\description{
score function
The score function completes all scoring for a single step of a scoresheet. operations are conducted in the following order:
\code{select}, \code{if_else}, \code{recode}, \code{sum}, \code{case_when}, \code{filter_at}
}
