% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vars.R
\name{select_vars}
\alias{select_vars}
\title{Select Variables Using Metadata}
\usage{
select_vars(raw, metadata)
}
\arguments{
\item{raw}{\itemize{
\item raw data object
}}

\item{metadata}{\itemize{
\item formatted metadata object
}}
}
\value{
tibble with selected variables
}
\description{
Select Variables Using Metadata
}
