% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_stats.R
\name{update_dnp}
\alias{update_dnp}
\title{Update DNP}
\usage{
update_dnp(game_stats_table, player_ids, game_id)
}
\arguments{
\item{game_stats_table}{Game stats table}

\item{player_ids}{Unique identifiers of a player from the players table}

\item{game_id}{Unique identifier of a game from the games table}
}
\description{
Update DNP (did not play) in game_stats_table
}
