% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{update_games_row}
\alias{update_games_row}
\title{Update games row}
\usage{
update_games_row(
  games_table,
  team_id,
  game_id,
  date,
  opponent,
  team_score,
  opponent_score
)
}
\arguments{
\item{games_table}{Games table}

\item{team_id}{Unique identifier of a team from the teams table}

\item{game_id}{Unique identifier of a game from the games table}

\item{date}{Date game was played}

\item{opponent}{Opponent of focal team}

\item{team_score}{Final score for focal team}

\item{opponent_score}{Final score for opponent}
}
\description{
Add or update row in games table
}
