% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binge_raw.R
\docType{data}
\name{binge_raw}
\alias{binge_raw}
\title{ALSPAC binge eating example dataset with 8 binge eating variables assessed}
\format{
R data file
}
\usage{
data(binge_raw)
}
\description{
A synthetic example dataset for binge eating items collected at age 14
in the ALSPAC cohort.
}
\examples{
data(binge_raw)
}
\keyword{datasets}
