% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_when_vars.R
\name{case_when_vars}
\alias{case_when_vars}
\title{Title}
\usage{
case_when_vars(raw, metadata)
}
\arguments{
\item{raw}{raw data object}

\item{metadata}{formatted metadata object}
}
\value{
a tibble with new variables appended that follow tidyverse evaluation of case_when metadata statements
}
\description{
Title
}
