% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_when_vars.R
\name{case_when_vars}
\alias{case_when_vars}
\title{case_when function}
\usage{
case_when_vars(raw, scoresheet)
}
\arguments{
\item{raw}{a raw data object}

\item{scoresheet}{a properly formatted scoresheet object. Key columns that must have data for a case_when function are new_var and code. Code should be formatted in accordance with
dplyr case_when input. See \href{https://dplyr.tidyverse.org/reference/case_when.html}{case_when}}
}
\value{
a tibble with new variables appended that follow tidyverse evaluation of case_when scoresheet statements
}
\description{
case_when function
}
