% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logs.R
\name{create_log_entry}
\alias{create_log_entry}
\title{Create log entry}
\usage{
create_log_entry(undo_bool, event = events, player)
}
\arguments{
\item{undo_bool}{Boolean indicating if event was undone}

\item{event}{Game event (e.g., FGM, STL, etc.)}

\item{player}{Player name}
}
\description{
Create entry for appending to game log
}
\examples{
create_log_entry(FALSE, "FGM3", "Stephen Curry")

}
