% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{create_player_names}
\alias{create_player_names}
\title{Create player names}
\usage{
create_player_names(
  first_names,
  last_names,
  numbers,
  tab = c("roster", "scorekeeper")
)
}
\arguments{
\item{first_names}{Vector of players' first names}

\item{last_names}{Vector of players' last names}

\item{numbers}{Vector of players' numbers}

\item{tab}{Tab where name will be displayed: roster or scorekeeper}
}
\description{
Create player names for display in app
}
