% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosters.R
\name{create_roster_view}
\alias{create_roster_view}
\title{Create roster view}
\usage{
create_roster_view(players_table, rosters_table)
}
\arguments{
\item{players_table}{Players table}

\item{rosters_table}{Rosters table}
}
\description{
Create roster view. A roster view is not saved to file.
}
