% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_at.R
\name{filter_at_vars}
\alias{filter_at_vars}
\title{filter_at_vars
\code{filter_at_vars} allows you to filter rows of a dataset based on values of specified columns}
\usage{
filter_at_vars(raw, scoresheet)
}
\arguments{
\item{raw}{a raw data object}

\item{scoresheet}{a scoresheet object. Key columns for the filter_at function are \code{code}. Code should include two arguments
separated by a comma. the first argument a numeric vector of column positions. *Note - this currently must be a numeric vector.
The second is a predicate function to be applied to the columns or a logical vector. Variables for which the predicate
returns true are selected. see dplyr for additional info. \href{https://dplyr.tidyverse.org/reference/filter_all.html}{filter_all}.}
}
\value{
tibble that is filtered at specific variables
}
\description{
filter_at_vars
\code{filter_at_vars} allows you to filter rows of a dataset based on values of specified columns
}
