% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_else_vars.R
\name{if_else_vars}
\alias{if_else_vars}
\title{Title}
\usage{
if_else_vars(raw, scoresheet)
}
\arguments{
\item{raw}{Raw data object}

\item{scoresheet}{Formatted scoresheet object}
}
\value{
a tibble with new variables appended that follow tidyverse evaluation of if_else scoresheet statements
}
\description{
Title
}
