% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-stats.R
\name{init_game_stats_table}
\alias{init_game_stats_table}
\title{Initialize game stats table}
\usage{
init_game_stats_table()
}
\description{
Initialize game stats table with no data. Only includes columns that are tallied, not calculated.
}
