% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_vars.R
\name{mean_vars}
\alias{mean_vars}
\title{Mean variables using scoresheet}
\usage{
mean_vars(raw, scoresheet)
}
\arguments{
\item{raw}{A raw data object}

\item{scoresheet}{A formatted scoresheet object}
}
\value{
raw tibble with four additional columns appended:
\code{{n_var}} : a column which is the mean of items, with NAs removed
\verb{\{n_var\}_complete} : a column with the mean of items -- rows that contain any NA values return NA
\verb{\{n_var\}_NAs} : a column with the count of NAs among items contributing to the mean
\verb{\{n_var\}_NA_percent} : a column with the percentage of NAs among the variables contributing to the mean
}
\description{
Mean variables using scoresheet
}
