% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_vars.R
\name{recode_vars}
\alias{recode_vars}
\title{The recode_vars function recodes variables according to metadata instructions}
\usage{
recode_vars(raw, metadata)
}
\arguments{
\item{raw}{: A raw data object}

\item{metadata}{: A metadata object}
}
\value{
A tibble with both original and recoded variables
}
\description{
The recode_vars function recodes variables according to metadata instructions
}
