% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_vars.R
\name{recode_vars}
\alias{recode_vars}
\title{recode_vars function
recodes variables according to scoresheet instructions}
\usage{
recode_vars(raw, scoresheet)
}
\arguments{
\item{raw}{: A raw data object}

\item{scoresheet}{: A scoresheet object}
}
\value{
A tibble with both original and recoded variables
}
\description{
recode_vars function
recodes variables according to scoresheet instructions
}
