% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorekeep.R
\name{score}
\alias{score}
\title{Title}
\usage{
score(raw, metadata)
}
\arguments{
\item{raw}{Raw data object}

\item{metadata}{Formatted metadata object with one step selected (all operations can be preformed simultaneously.
For sequential operations, see scorekeep function)}
}
\value{
a tibble
}
\description{
Title
}
