% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorekeep.R
\name{score}
\alias{score}
\title{score function
The score function completes all scoring for a single step of a scoresheet. Operations are conducted in the following order:
\code{select}, \code{if_else}, \code{recode}, \code{sum}, \code{case_when}, \code{filter_at}, \code{rename}, \code{mean}}
\usage{
score(raw, scoresheet)
}
\arguments{
\item{raw}{A raw data object}

\item{scoresheet}{A formatted scoresheet object with one step selected, operations that can all be preformed
in a single step, simultaneously
For sequential operations, see scorekeep function)}
}
\value{
a tibble
}
\description{
score function
The score function completes all scoring for a single step of a scoresheet. Operations are conducted in the following order:
\code{select}, \code{if_else}, \code{recode}, \code{sum}, \code{case_when}, \code{filter_at}, \code{rename}, \code{mean}
}
