% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorekeep.R
\name{scorekeep}
\alias{scorekeep}
\title{Scores raw data in a step-by-step fashion using metadata provided}
\usage{
scorekeep(raw, metadata)
}
\arguments{
\item{raw}{Raw data object}

\item{metadata}{Formatted metadata object}
}
\value{
a series of tibbles representing data manipulations at each step as specified in metadata file
}
\description{
Scores raw data in a step-by-step fashion using metadata provided
}
\examples{

}
