% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vars.R
\name{select_vars}
\alias{select_vars}
\title{Select Variables Using scoresheet}
\usage{
select_vars(raw, scoresheet)
}
\arguments{
\item{raw}{\itemize{
\item raw data object
}}

\item{scoresheet}{\itemize{
\item formatted scoresheet object
}}
}
\value{
tibble with selected variables
}
\description{
Select Variables Using scoresheet
}
