% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_vars.R
\name{sum_vars}
\alias{sum_vars}
\title{Sum variables using scoresheet}
\usage{
sum_vars(raw, scoresheet)
}
\arguments{
\item{raw}{A raw data object}

\item{scoresheet}{A formatted scoresheet object}
}
\value{
raw tibble with four additional columns appended:
\code{{n_var}} : a column which is the sum of items - NAs are treated as '0' values
\verb{\{n_var\}_complete} : a column with the sum of items -- rows that contain any NA values return NA
\verb{\{n_var\}_NAs} : a column with the count of NAs among items contributing to the sum
\verb{\{n_var\}_NA_percent} : a column with the percentage of NAs among the variables contributing to the sum
\verb{\{n_var\}_weighted_sum} : a column with a sum weighted by the number of items contributing to the sum
}
\description{
Sum variables using scoresheet
}
