% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-stats.R
\name{update_game_stat}
\alias{update_game_stat}
\title{Update game stat}
\usage{
update_game_stat(
  games_stats_table,
  player_id,
  game_id,
  stat = events,
  undo = FALSE
)
}
\arguments{
\item{player_id}{Unique identifier of a player from the players table}

\item{game_id}{Unique identifier of a game from the games table}

\item{stat}{Statistic (column) to be changed}

\item{undo}{When true, stat is decremented.}

\item{game_stats_table}{Game stats table}
}
\description{
Update game stat in game_stats_table
}
